requirements.txt
fastapi
mangum
python-dotenv
google-cloud-bigquery